removed some debugging code from main, updated the README in titleScreen and removed uneeded backups.
