+++ src/titleScreen/README.md
@@ -1,0 +1,12 @@
+<h1>TitleScreen Class</h1>
+<p>
+The only function of this cluass will be to bring up the title screen
+</p>
+<p>
+This is not supposed to bring up the main menu or even start the game.<br> 
+I want the calling functionto handle that
+</p>
+<h2>TODO</h2>
+<p>- boil the title screen class down to only the title screen</p>
+<p>- seperate the data from the functions</p>
+<p>- prepare the class for the data to be loaded from files for when that does happen</p>
