--- src/titleScreenNew/README.md
@@ -1,7 +1,0 @@
-<h1>TitleScreen Class</h1>
-<p>
-The only function of this class will be to bring up the title screen
-</p>
-<p>
-This is not supposed to bring up the main menu or even start the game.<br> 
-I want the calling functionto handle that\
